+++ packages/apps/bridges/helix/substrate-dvm/utils/bridge.ts
@@ -11,7 +11,7 @@ import { convertToSS58, dvmAddressToAccountId } from 'shared/utils/helper/addres
 import { toWei } from 'shared/utils/helper/balance';
 import { typeRegistryFactory } from 'shared/utils/helper/huge';
 import { isNativeToken, isRing } from 'shared/utils/helper/validator';
-import { genEthereumTransactionObs, signAndSendExtrinsic } from 'shared/utils/tx';
+import { sendTransaction, signAndSendExtrinsic } from 'shared/utils/tx';
 import { Bridge, TokenWithAmount } from '../../../../core/bridge';
 import { IssuingPayload, RedeemPayload, SubstrateDVMBridgeConfig } from '../model';
 
@@ -78,7 +78,7 @@ export class SubstrateDVMBridge extends Bridge<SubstrateDVMBridgeConfig, Polkado
               })
             ).pipe(
               switchMap((gas) =>
-                genEthereumTransactionObs({
+                sendTransaction({
                   from: sender,
                   to: SUBSTRATE_DVM_WITHDRAW,
                   data: u8aToHex(extrinsic.method.toU8a()),
